process_financial/requirements.txt
gspread>=5.12.0
gspread_dataframe>=3.3.1
oauth2client>=4.1.3
google-cloud-secret-manager>=2.16.2
google-cloud-storage>=2.16.0
pandas>=2.2.2
pyarrow>=15.0.2
